+++ sql/schema/007_problem.sql
@@ -6,7 +6,7 @@ CREATE TABLE problem(
   description TEXT NOT NULL,
   title VARCHAR(255) NOT NULL,
   memory_limit INTEGER NOT NULL,
-  time_limit DOUBLE PRECISION NOT NULL,
+  time_limit INT NOT NULL,
   quiz_id CHAR(36) NOT NULL,
   FOREIGN KEY (quiz_id) REFERENCES quiz(id) ON DELETE CASCADE
 );
